fix: moved SmsWebhooks to Hooks namespace
