+++ src/Sinch/SMS/Hooks/IInboundMessage.cs
@@ -3,7 +3,7 @@ namespace Sinch.SMS.Hooks
     /// <summary>
     ///     Base interface for all inbound message types.
     /// </summary>
-    /// <seealso cref="Sinch.SMS.Webhooks.ISmsWebhooks.ParseEvent"/>
+    /// <seealso cref="ISmsWebhooks.ParseEvent"/>
     public interface IInboundMessage : ISmsEvent
     {
     }
