+++ src/Sinch/SMS/Hooks/ISmsEvent.cs
@@ -5,7 +5,7 @@ namespace Sinch.SMS.Hooks
     /// <summary>
     ///     Base class for all WebHook event's class
     /// </summary>
-    /// <seealso cref="Sinch.SMS.Webhooks.SmsWebhooks"/>
+    /// <seealso cref="SmsWebhooks"/>
     [JsonConverter(typeof(SmsEventConverter))]
     public interface ISmsEvent
     {
